simx86: remove no longer needed slow path
When generating code the existing code will be deleted anyway
in DoClose(). But to not have a performance regression it does mean we
need to check for correctly placed existing code, via FindTree
to see if compilation should end, otherwise it will just keep
going and overwrite (unless prejit, to respect its gap).